+++ ibm_boto3/docs/docstring.py
@@ -12,15 +12,18 @@
 # language governing permissions and limitations under the License.
 from ibm_botocore.docs.docstring import LazyLoadedDocstring
 
-from ibm_boto3.docs.action import document_action
-from ibm_boto3.docs.action import document_load_reload_action
+from ibm_boto3.docs.action import document_action, document_load_reload_action
+from ibm_boto3.docs.attr import (
+    document_attribute,
+    document_identifier,
+    document_reference,
+)
+from ibm_boto3.docs.collection import (
+    document_batch_action,
+    document_collection_method,
+    document_collection_object,
+)
 from ibm_boto3.docs.subresource import document_sub_resource
-from ibm_boto3.docs.attr import document_attribute
-from ibm_boto3.docs.attr import document_identifier
-from ibm_boto3.docs.attr import document_reference
-from ibm_boto3.docs.collection import document_collection_object
-from ibm_boto3.docs.collection import document_collection_method
-from ibm_boto3.docs.collection import document_batch_action
 from ibm_boto3.docs.waiter import document_resource_waiter