+++ ibm_boto3/session.py
@@ -35,7 +35,7 @@ from ibm_boto3.exceptions import ResourceNotExistsError, UnknownAPIVersionError
 from .resources.factory import ResourceFactory
 
 
-class Session(object):
+class Session:
     """
     A session stores configuration state and allows you to create service
     clients and resources.
@@ -78,11 +78,20 @@ class Session(object):
     :param profile_name: The name of a profile to use. If not given, then
                          the default profile is used.
     """
-    def __init__(self, aws_access_key_id=None, aws_secret_access_key=None,
-                 aws_session_token=None, region_name=None,
-                 ibm_api_key_id=None, ibm_service_instance_id=None, ibm_auth_endpoint=None,
-                 auth_function=None, token_manager=None,
-                 botocore_session=None, profile_name=None):
+    def __init__(
+        self,
+        aws_access_key_id=None,
+        aws_secret_access_key=None,
+        aws_session_token=None,
+        region_name=None,
+        ibm_api_key_id=None,
+        ibm_service_instance_id=None,
+        ibm_auth_endpoint=None,
+        auth_function=None,
+        token_manager=None,
+        botocore_session=None,
+        profile_name=None,
+    ):
         if botocore_session is not None:
             self._session = botocore_session
         else:
@@ -91,8 +100,9 @@ class Session(object):
 
         # Setup custom user-agent string if it isn't already customized
         if self._session.user_agent_name == 'ibm-cos-sdk-python-core':
-            botocore_info = 'ibm-cos-sdk-python-core/{0}'.format(
-                self._session.user_agent_version)
+            botocore_info = 'ibm-cos-sdk-python-core/{}'.format(
+                self._session.user_agent_version
+            )
             if self._session.user_agent_extra:
                 self._session.user_agent_extra += ' ' + botocore_info
             else:
@@ -108,20 +118,23 @@ class Session(object):
                 access_key=aws_access_key_id, secret_key=aws_secret_access_key, token=aws_session_token,
                 ibm_api_key_id=ibm_api_key_id, ibm_service_instance_id=ibm_service_instance_id,
                 ibm_auth_endpoint=ibm_auth_endpoint,
-                auth_function=auth_function, token_manager=token_manager)
+                auth_function=auth_function, token_manager=token_manager
+            )
 
         if region_name is not None:
             self._session.set_config_variable('region', region_name)
 
         self.resource_factory = ResourceFactory(
-            self._session.get_component('event_emitter'))
+            self._session.get_component('event_emitter')
+        )
         self._setup_loader()
         self._register_default_handlers()
 
     def __repr__(self):
-        return '{0}(region_name={1})'.format(
+        return '{}(region_name={})'.format(
             self.__class__.__name__,
-            repr(self._session.get_config_variable('region')))
+            repr(self._session.get_config_variable('region')),
+        )
 
     @property
     def profile_name(self):
@@ -157,7 +170,8 @@ class Session(object):
         """
         self._loader = self._session.get_component('data_loader')
         self._loader.search_paths.append(
-            os.path.join(os.path.dirname(__file__), 'data'))
+            os.path.join(os.path.dirname(__file__), 'data')
+        )
 
     def get_available_services(self):
         """
@@ -187,8 +201,9 @@ class Session(object):
         """
         return self._session.get_available_partitions()
 
-    def get_available_regions(self, service_name, partition_name='aws',
-                              allow_non_regional=False):
+    def get_available_regions(
+        self, service_name, partition_name='aws', allow_non_regional=False
+    ):
         """Lists the region and endpoint names of a particular partition.
 
         :type service_name: string
@@ -207,12 +222,14 @@ class Session(object):
         :return: Returns a list of endpoint names (e.g., ["us-east-1"]).
         """
         return self._session.get_available_regions(
-            service_name=service_name, partition_name=partition_name,
-            allow_non_regional=allow_non_regional)
+            service_name=service_name,
+            partition_name=partition_name,
+            allow_non_regional=allow_non_regional,
+        )
 
     def get_credentials(self):
         """
-        Return the :class:`ibm_botocore.credential.Credential` object
+        Return the :class:`ibm_botocore.credentials.Credentials` object
         associated with this session.  If the credentials have not
         yet been loaded, this will attempt to load them.  If they
         have already been loaded, this will return the cached
@@ -220,12 +237,36 @@ class Session(object):
         """
         return self._session.get_credentials()
 
-    def client(self, service_name, region_name=None, api_version=None,
-               use_ssl=True, verify=None, endpoint_url=None,
-               aws_access_key_id=None, aws_secret_access_key=None, aws_session_token=None,
-               ibm_api_key_id=None, ibm_service_instance_id=None, ibm_auth_endpoint=None,
-               auth_function=None, token_manager=None,
-               config=None):
+    def get_partition_for_region(self, region_name):
+        """Lists the partition name of a particular region.
+
+        :type region_name: string
+        :param region_name: Name of the region to list partition for (e.g.,
+             us-east-1).
+
+        :rtype: string
+        :return: Returns the respective partition name (e.g., aws).
+        """
+        return self._session.get_partition_for_region(region_name)
+
+    def client(
+        self,
+        service_name,
+        region_name=None,
+        api_version=None,
+        use_ssl=True,
+        verify=None,
+        endpoint_url=None,
+        aws_access_key_id=None,
+        aws_secret_access_key=None,
+        aws_session_token=None,
+        ibm_api_key_id=None,
+        ibm_service_instance_id=None,
+        ibm_auth_endpoint=None,
+        auth_function=None,
+        token_manager=None,
+        config=None,
+    ):
         """
         Create a low-level service client by name.
 
@@ -314,21 +355,41 @@ class Session(object):
 
         """
         return self._session.create_client(
-            service_name, region_name=region_name, api_version=api_version,
-            use_ssl=use_ssl, verify=verify, endpoint_url=endpoint_url,
+            service_name,
+            region_name=region_name,
+            api_version=api_version,
+            use_ssl=use_ssl,
+            verify=verify,
+            endpoint_url=endpoint_url,
             aws_access_key_id=aws_access_key_id,
             aws_secret_access_key=aws_secret_access_key,
             aws_session_token=aws_session_token,
-            ibm_api_key_id=ibm_api_key_id, ibm_service_instance_id=ibm_service_instance_id,
-            ibm_auth_endpoint=ibm_auth_endpoint, auth_function=auth_function,
-            token_manager=token_manager, config=config)
+            ibm_api_key_id=ibm_api_key_id,
+            ibm_service_instance_id=ibm_service_instance_id,
+            ibm_auth_endpoint=ibm_auth_endpoint,
+            auth_function=auth_function,
+            token_manager=token_manager,
+            config=config,
+        )
 
-    def resource(self, service_name, region_name=None, api_version=None,
-                 use_ssl=True, verify=None, endpoint_url=None,
-                 aws_access_key_id=None, aws_secret_access_key=None, aws_session_token=None,
-                 ibm_api_key_id=None, ibm_service_instance_id=None, ibm_auth_endpoint=None,
-                 auth_function=None, token_manager=None,
-                 config=None):
+    def resource(
+        self,
+        service_name,
+        region_name=None,
+        api_version=None,
+        use_ssl=True,
+        verify=None,
+        endpoint_url=None,
+        aws_access_key_id=None,
+        aws_secret_access_key=None,
+        aws_session_token=None,
+        ibm_api_key_id=None,
+        ibm_service_instance_id=None,
+        ibm_auth_endpoint=None,
+        auth_function=None,
+        token_manager=None,
+        config=None,
+    ):
         """
         Create a resource service client by name.
 
@@ -419,19 +480,24 @@ class Session(object):
         """
         try:
             resource_model = self._loader.load_service_model(
-                service_name, 'resources-1', api_version)
+                service_name, 'resources-1', api_version
+            )
         except UnknownServiceError:
             available = self.get_available_resources()
             has_low_level_client = (
-                service_name in self.get_available_services())
-            raise ResourceNotExistsError(service_name, available,
-                                         has_low_level_client)
+                service_name in self.get_available_services()
+            )
+            raise ResourceNotExistsError(
+                service_name, available, has_low_level_client
+            )
         except DataNotFoundError:
             # This is because we've provided an invalid API version.
             available_api_versions = self._loader.list_api_versions(
-                service_name, 'resources-1')
+                service_name, 'resources-1'
+            )
             raise UnknownAPIVersionError(
-                service_name, api_version, ', '.join(available_api_versions))
+                service_name, api_version, ', '.join(available_api_versions)
+            )
 
         if api_version is None:
             # Even though ibm_botocore's load_service_model() can handle
@@ -450,7 +516,8 @@ class Session(object):
             # and loader.determine_latest_version(..., 'resources-1')
             # both load the same api version of the file.
             api_version = self._loader.determine_latest_version(
-                service_name, 'resources-1')
+                service_name, 'resources-1'
+            )
 
         # Creating a new resource instance requires the low-level client
         # and service model, the resource version and resource JSON data.
@@ -463,8 +530,12 @@ class Session(object):
         else:
             config = Config(user_agent_extra='Resource')
         client = self.client(
-            service_name, region_name=region_name, api_version=api_version,
-            use_ssl=use_ssl, verify=verify, endpoint_url=endpoint_url,
+            service_name,
+            region_name=region_name,
+            api_version=api_version,
+            use_ssl=use_ssl,
+            verify=verify,
+            endpoint_url=endpoint_url,
             aws_access_key_id=aws_access_key_id,
             aws_secret_access_key=aws_secret_access_key,
             aws_session_token=aws_session_token,
@@ -473,24 +544,26 @@ class Session(object):
             ibm_auth_endpoint=ibm_auth_endpoint,
             auth_function=auth_function,
             token_manager=token_manager,
-            config=config)
+            config=config,
+        )
         service_model = client.meta.service_model
 
         # Create a ServiceContext object to serve as a reference to
         # important read-only information about the general service.
         service_context = ibm_boto3.utils.ServiceContext(
-            service_name=service_name, service_model=service_model,
+            service_name=service_name,
+            service_model=service_model,
             resource_json_definitions=resource_model['resources'],
             service_waiter_model=ibm_boto3.utils.LazyLoadedWaiterModel(
                 self._session, service_name, api_version
-            )
+            ),
         )
 
         # Create the service resource class.
         cls = self.resource_factory.load_from_definition(
             resource_name=service_name,
             single_resource_json_definition=resource_model['service'],
-            service_context=service_context
+            service_context=service_context,
         )
 
         return cls(client=client)
@@ -501,15 +574,17 @@ class Session(object):
         self._session.register(
             'creating-client-class.s3',
             ibm_boto3.utils.lazy_call(
-                'ibm_boto3.s3.inject.inject_s3_transfer_methods'))
+                'ibm_boto3.s3.inject.inject_s3_transfer_methods'
+            ),
+        )
         self._session.register(
             'creating-resource-class.s3.Bucket',
-            ibm_boto3.utils.lazy_call(
-                'ibm_boto3.s3.inject.inject_bucket_methods'))
+            ibm_boto3.utils.lazy_call('ibm_boto3.s3.inject.inject_bucket_methods'),
+        )
         self._session.register(
             'creating-resource-class.s3.Object',
-            ibm_boto3.utils.lazy_call(
-                'ibm_boto3.s3.inject.inject_object_methods'))
+            ibm_boto3.utils.lazy_call('ibm_boto3.s3.inject.inject_object_methods'),
+        )
         self._session.register(
             'creating-resource-class.s3.ObjectSummary',
             ibm_boto3.utils.lazy_call(