+++ tests/functional/test_resource.py
@@ -10,10 +10,11 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
+import ibm_botocore.session
+import pytest
+
 import ibm_boto3
 from ibm_boto3.exceptions import ResourceNotExistsError
-
-import ibm_botocore.session
 from tests import unittest
 
 
@@ -28,28 +29,28 @@ class TestResourceCustomization(unittest.TestCase):
     def add_new_method(self, name):
         def handler(class_attributes, **kwargs):
             class_attributes[name] = identity
+
         return handler
 
     def test_can_inject_method_onto_resource(self):
         session = ibm_boto3.Session(botocore_session=self.botocore_session)
-        self.botocore_session.register('creating-resource-class.s3',
-                                       self.add_new_method(name='my_method'))
+        self.botocore_session.register(
+            'creating-resource-class.s3', self.add_new_method(name='my_method')
+        )
         resource = session.resource('s3')
-        self.assertTrue(hasattr(resource, 'my_method'))
-        self.assertEqual(resource.my_method('anything'), 'anything')
+        assert hasattr(resource, 'my_method')
+        assert resource.my_method('anything') == 'anything'
 
 
 class TestSessionErrorMessages(unittest.TestCase):
     def test_has_good_error_message_when_no_resource(self):
         bad_resource_name = 'doesnotexist'
-        err_regex = (
-            '%s.*resource does not exist.' % bad_resource_name
-        )
-        with self.assertRaisesRegex(ResourceNotExistsError, err_regex):
+        err_regex = f'{bad_resource_name}.*resource does not exist.'
+        with pytest.raises(ResourceNotExistsError, match=err_regex):
             ibm_boto3.resource(bad_resource_name)
 
 
 class TestGetAvailableSubresources(unittest.TestCase):
     def test_s3_available_subresources_exists(self):
         s3 = ibm_boto3.resource('s3')
-        self.assertTrue(hasattr(s3, 'get_available_subresources'))
+        assert hasattr(s3, 'get_available_subresources')