+++ tests/unit/docs/test_utils.py
@@ -10,33 +10,32 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-from tests import unittest
-
-from ibm_boto3.resources.model import Parameter
 from ibm_boto3.docs.utils import get_resource_ignore_params
+from ibm_boto3.resources.model import Parameter
+from tests import unittest
 
 
 class TestGetResourceIgnoreParams(unittest.TestCase):
     def test_target_is_single_resource(self):
         param = Parameter('InstanceId', 'response')
         ignore_params = get_resource_ignore_params([param])
-        self.assertEqual(ignore_params, ['InstanceId'])
+        assert ignore_params == ['InstanceId']
 
     def test_target_is_multiple_resources(self):
         param = Parameter('InstanceIds[]', 'response')
         ignore_params = get_resource_ignore_params([param])
-        self.assertEqual(ignore_params, ['InstanceIds'])
+        assert ignore_params == ['InstanceIds']
 
     def test_target_is_element_of_multiple_resources(self):
         param = Parameter('InstanceIds[0]', 'response')
         ignore_params = get_resource_ignore_params([param])
-        self.assertEqual(ignore_params, ['InstanceIds'])
+        assert ignore_params == ['InstanceIds']
 
     def test_target_is_nested_param(self):
         param = Parameter('Filters[0].Name', 'response')
         ignore_params = get_resource_ignore_params([param])
-        self.assertEqual(ignore_params, ['Filters'])
+        assert ignore_params == ['Filters']
 
         param = Parameter('Filters[0].Values[0]', 'response')
         ignore_params = get_resource_ignore_params([param])
-        self.assertEqual(ignore_params, ['Filters'])
+        assert ignore_params == ['Filters']